requirements.txt
streamlit
joblib
tldextract
python-whois
dnspython
requests
scikit-learn